Removed EuiLoadingTableBody & moved its logic into EuiBasicTable to prevent re-rendering the whole table when loading is set to true
