--- src/components/basic_table/loading_table_body.test.tsx
@@ -1,20 +1,0 @@
-import React from 'react';
-import { mount } from 'enzyme';
-import { requiredProps } from '../../test';
-
-import { LoadingTableBody } from './loading_table_body';
-
-describe('LoadingTableBody', () => {
-  test('render', () => {
-    const props = {
-      ...requiredProps,
-    };
-    const component = mount(
-      <table>
-        <LoadingTableBody {...props} />
-      </table>
-    );
-
-    expect(component).toMatchSnapshot();
-  });
-});
